[#16] 날짜 계산 부분 Date extension으로 분리
